DSN-028 (phase 3): move auth + catalog to internal/
Third phase of the DSN-028 6-phase package restructure. Smallest move
phase: just auth and the outbound catalog client.

Moves
- core/auth/jwt.go -> internal/auth/jwt.go
- core/auth/jwt_test.go -> internal/auth/jwt_test.go
- api/authapi.go -> internal/auth/transport_http.go
- core/catalog/client.go -> internal/catalog/client.go
- core/catalog/client_test.go -> internal/catalog/client_test.go

After this phase: core/ contains only the platform subdirs (cache,
observability, ratelimit, secrets) and core/repo.go. api/ contains
only middleware, router glue, health, env, and the auth integration
tests.

Conventions extended
- jwt.go kept as the source filename (descriptive of the auth
  package's single concern; no equivalent to Phase 1's service.go
  convention since there's no Service struct here, just a Signer).
- @name annotation on TokenResponse preserves the short OpenAPI
  schema name.
- basicAuthErr (an unexported 4-line helper from api/middleware.go)
  inlined into internal/auth/transport_http.go where it has its only
  caller. The orphan removed from middleware.go.

Tests intentionally kept in api/
- api/authapi_test.go and api/auth_metrics_test.go stay in api/.
  They exercise the full api.ConfigureRouter behavior end-to-end
  (Authenticate middleware, AdminOnly, /metrics endpoint, basic-auth
  rejection) rather than internal/auth in isolation. The handler
  unit tests live with the source in internal/auth; these are
  router-level integration tests of api/'s auth wiring.

Out of scope / deferred to Phase 4
- api/middleware.go (Authenticate, AdminOnly, CtxKeyUser, Metrics,
  Logging, CorrelationLogger) stays in api/ for now. Moves to
  internal/platform/httpx/ in Phase 4 along with the rest of api/'s
  remaining cross-cutting code.

Acceptance criteria addressed
- core/auth, core/catalog no longer exist
- api/authapi.go removed
- go build ./..., make verify pass
- DSN-015 demo orchestrator continues to pass

Diff size: 17 files changed, 63 insertions / 63 deletions with
--find-renames. Well under the 600 LOC stop-line.
