+++ cmd/main.go
@@ -30,9 +30,7 @@ import (
 	"github.com/rs/zerolog/log"
 	"github.com/sksmith/go-micro-example/api"
 	"github.com/sksmith/go-micro-example/config"
-	"github.com/sksmith/go-micro-example/core/auth"
 	"github.com/sksmith/go-micro-example/core/cache"
-	"github.com/sksmith/go-micro-example/core/catalog"
 	"github.com/sksmith/go-micro-example/core/observability"
 	"github.com/sksmith/go-micro-example/core/ratelimit"
 	"github.com/sksmith/go-micro-example/core/secrets"
@@ -40,6 +38,8 @@ import (
 	"github.com/sksmith/go-micro-example/events"
 	"github.com/sksmith/go-micro-example/idempotency"
 	restidempotency "github.com/sksmith/go-micro-example/idempotency/rest"
+	"github.com/sksmith/go-micro-example/internal/auth"
+	"github.com/sksmith/go-micro-example/internal/catalog"
 	"github.com/sksmith/go-micro-example/internal/inventory"
 	"github.com/sksmith/go-micro-example/internal/user"
 	gmekafka "github.com/sksmith/go-micro-example/kafka"
