+++ internal/catalog/client.go
