+++ internal/catalog/client_test.go
@@ -10,8 +10,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/sksmith/go-micro-example/core/catalog"
 	"github.com/sksmith/go-micro-example/core/observability"
+	"github.com/sksmith/go-micro-example/internal/catalog"
 )
 
 func newClient(t *testing.T, server *httptest.Server, cfg catalog.Config) *catalog.HTTPClient {
